Add History View.

+++ TravelRefunds/Models/TravelQuery.cs
@@ -5,7 +5,7 @@
         public string From { get; set; }
         public string To { get; set; }
         public DistanceUnit DistanceUnit { get; set; }
-        public int? Distance { get; set; }
+        public double? Distance { get; set; }
     };
 
     public enum DistanceUnit
